consumer/requirements.txt
kafka-python
pymongo
pyyaml